\name{@PKGNAME@}
\docType{data}
\alias{@PKGNAME@}
\title{Greek mythology package: @TITLE@}
\description{Greek mythology package: @TITLE@}
\usage{
data(@PKGNAME@)
}
\format{@FORMAT@}
\details{According to @AUTHOR@'s Theogony, Zeus lay with Mnemosyne
  for nine days, and she gave birth to the Muses, who rejoice in their
  bright dancing places on Mount Helicon.}
\seealso{\code{\link[@PKGNAME@:get@PKGNAME@]{get@PKGNAME@}}}
\examples{
  data(@PKGNAME@)
  @PKGNAME@
}
\keyword{datasets}

