% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_point_forecasts.R
\name{add_point_forecasts}
\alias{add_point_forecasts}
\title{Add point forecasts to a quantile forecasts data.frame}
\usage{
add_point_forecasts(data)
}
\arguments{
\item{data}{The \code{data.frame} containing quantile forecasts}
}
\value{
The \code{data.frame} from \code{data} with additional rows containing the point
forecasts
}
\description{
Add point forecasts to a quantile forecasts data.frame
}
