% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegImpute.R
\name{backfill}
\alias{backfill}
\title{Fill-in missing values using simple imputation.}
\usage{
backfill(data, fillmethod = fillmethod)
}
\arguments{
\item{data}{A matrix of values.}

\item{fillmethod}{A string identifying the method to be used.}
}
\value{
The imputed matrix of \code{data}.
}
\description{
Fill-in missing values using simple imputation.
}
