% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_baseline.R
\name{build_baseline}
\alias{build_baseline}
\title{Build quantile baseline based on \code{\link[covidModels:fit_quantile_baseline]{covidModels::fit_quantile_baseline()}}}
\usage{
build_baseline(inc_obs, quantiles, horizon)
}
\arguments{
\item{inc_obs}{Observed number of incident cases}

\item{quantiles}{Vector of numeric values determining which quantiles will
be returned in the prediction}

\item{horizon}{Integer value determining the time horizon until which the
predictions will be made}
}
\description{
Build quantile baseline based on \code{\link[covidModels:fit_quantile_baseline]{covidModels::fit_quantile_baseline()}}
}
