% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cOde.R
\name{compileAndLoad}
\alias{compileAndLoad}
\title{Compile and load shared object implementing the ODE system.}
\usage{
compileAndLoad(modelname, filename, fcontrol, verbose)
}
\arguments{
\item{modelname}{Base name for source and dll file.}

\item{filename}{Full file name of the source file.}

\item{fcontrol}{Interpolation method for forcings.}

\item{verbose}{Print compiler output or not.}
}
\description{
Compile and load shared object implementing the ODE system.
}
\author{
Daniel Kaschek, \email{daniel.kaschek@physik.uni-freiburg.de}

Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

