% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ensemble_average.R
\name{create_ensemble_average}
\alias{create_ensemble_average}
\title{Create ensemble using mean or median average.}
\usage{
create_ensemble_average(forecasts, method = c("mean", "median"))
}
\arguments{
\item{forecasts}{\code{data.frame} containing all the forecasts to be summarised
as an ensemble.}

\item{method}{One of \code{mean} (default) or \code{median}.}
}
\value{
ensemble model
}
\description{
Create ensemble using mean or median average.
}
