% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ensemble_relative_skill.R
\name{create_ensemble_relative_skill}
\alias{create_ensemble_relative_skill}
\title{Ensemble by relative skill}
\usage{
create_ensemble_relative_skill(
  forecasts,
  evaluation_date,
  continuous_weeks = 4,
  average = "mean",
  by_horizon = FALSE,
  return_criteria = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{forecasts}{\code{data.frame} containing all the forecasts to be summarised
as an ensemble.}

\item{continuous_weeks}{include only forecasts with a history of evaluation}

\item{average}{Method to average. One of \code{"mean"} or \code{"median"}}

\item{by_horizon}{weight using relative skill by horizon, rather than average}

\item{verbose}{Logical determining whether diagnostic messages should
be printed while running (defaults to \code{FALSE}).}

\item{forecast_date}{date with saved evaluation csv of forecasts}
}
\description{
Ensemble by relative skill
}
\details{
Ensemble = sum of forecast values weighted by the inverse of relative skill
Weights are by model, horizon, target, location
i.e. not weighted by quantile
}
