% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cOde.R
\name{cvodesSyntax}
\alias{cvodesSyntax}
\title{Sundials::cvodes: Replace powers and symbols to get correct C++ syntax for
cvodes.}
\usage{
cvodesSyntax(f, variables, parameters, varSym = "y", parSym = "p",
  forcings = NULL)
}
\arguments{
\item{f}{Named character vector.}

\item{variables}{Variables appearing on the ODE, \code{names(f)}.}

\item{parameters}{Parameters appearing in the ODE.}

\item{varSym}{Symbol of the variables to appear in the source code.}

\item{parSym}{Symbol of the parameter to appear in the source code.}

\item{forcings}{Inhomogenity of the ODE. Currently not used.}
}
\description{
Sundials::cvodes: Replace powers and symbols to get correct C++ syntax for
cvodes.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

