% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_1.R
\name{density_function}
\alias{density_function}
\title{Bind two factors}
\usage{
density_function(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
