% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhts.R
\name{dhts}
\alias{dhts}
\title{discrete hierarchical time series constructor.}
\usage{
dhts(bts, s_mat, domain)
}
\arguments{
\item{bts}{bottom-level time series, T * m}

\item{s_mat}{the summing matrix}

\item{domain}{the domain of bottom-level series, 2 * m matrix, 
Each column corresponds to each series, and the first row corresponds to
the minimums, and the second row corresponds the maximums.}
}
\value{
a dhts object
}
\description{
Create a 'dhts' object
}
\examples{
# create a 3-nodes discrete hts.
bts <- matrix(sample(0:1, 100, replace=TRUE)
s_mat <- matrix(c(1, 1, 0, 1, 0, 1), 3)
domain <- matrix(c(1, 0, 1, 0), 2)
dhts(bts, s_mat, domain)
}
