% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algorithm Implementation}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Data Frame of Graph}

\item{init_node}{Numeric Scalar Input}
}
\value{
Return a Numeric Vector with the shortest path between nodes in the graph.
}
\description{
Dijkstra Algorithm to find the shortest path between nodes in the graph giving a shortest path vector.
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
