% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DreamAI_bagging.R
\name{DreamAI_Bagging}
\alias{DreamAI_Bagging}
\title{Bag Imputation of Missing Protein Abundances with Iterative Prediction Model}
\usage{
DreamAI_Bagging(data, k = 10, maxiter_MF = 10, ntree = 100,
  maxnodes = NULL, maxiter_ADMIN = 30, tol = 10^(-2),
  gamma_ADMIN = NA, gamma = 50, CV = FALSE,
  fillmethod = "row_mean", maxiter_RegImpute = 10,
  conv_nrmse = 1e-06, iter_SpectroFM = 40, method = c("KNN",
  "MissForest", "ADMIN", "Brinn", "SpectroFM", "RegImpute", "Ensemble"),
  SamplesPerBatch, n.bag, save.out = TRUE, path = NULL, ProcessNum)
}
\arguments{
\item{data}{dataset in the form of a matrix or dataframe with missing values or NA's. The function throws an error message and stops if any row or column in the dataset is missing all values}

\item{k}{number of neighbors to be used in the imputation by KNN and ADMIN (default is 10)}

\item{maxiter_MF}{maximum number of iteration to be performed in the imputation by "MissForest" if the stopping criteria is not met beforehand}

\item{ntree}{number of trees to grow in each forest in "MissForest"}

\item{maxnodes}{maximum number of terminal nodes for trees in the forest in "MissForest", has to equal at least the number of columns in the given data}

\item{maxiter_ADMIN}{maximum number of iteration to be performed in the imputation by "ADMIN" if the stopping criteria is not met beforehand}

\item{tol}{convergence threshold for "ADMIN"}

\item{gamma_ADMIN}{parameter for ADMIN to control abundance dependent missing. Set gamma_ADMIN=0 for log ratio intensity data. For abundance data put gamma_ADMIN=NA, and it will be estimated accordingly}

\item{gamma}{parameter of the supergradients of popular nonconvex surrogate functions, e.g. SCAD and MCP of L0-norm for Brinn}

\item{CV}{a logical value indicating whether to fit the best gamma with cross validation for "Brinn". If CV=FALSE, default gamma=50 is used, while if CV=TRUE gamma is calculated using cross-validation.}

\item{fillmethod}{a string identifying the method to be used to initially filling the missing values using simple imputation for "RegImpute". That could be "row_mean" or "zeros", with "row_mean" being the default. It throws an warning if "row_median" is used.}

\item{maxiter_RegImpute}{maximum number of iterations to reach convergence in the imputation by "RegImpute"}

\item{conv_nrmse}{convergence threshold for "RegImpute"}

\item{iter_SpectroFM}{number of iterations for "SpectroFM"}

\item{method}{a vector of imputation methods: ("KNN", "MissForest", "ADMIN", "Brinn", "SpectroFM, "RegImpute", "Ensemble"). Default is "Ensemble" if nothing is specified}

\item{SamplesPerBatch}{number of samples per batch (batch size in the original data)}

\item{n.bag}{number of pseudo datasets to generate and impute in the current process
save.out logical indicator whether or not to save the output. When TRUE output is saved, when FALSE output is returned}

\item{path}{location to save the output file from the curent process.  Path only needs to be specified when save.out=TRUE}

\item{ProcessNum}{process number starting from 1 when run in cluster, e.g. 1 - 10, 1 - 100 etc. Needs to be specified only if the output is saved}
}
\value{
list of imputed dataset, n.bag and a matrix containing gene name, sample name, true and imputed values of every pseudo missing combined from n.bag datasets. Impute is a list of imputed datasets (average over all pseudo imputed data matrices) by different methods as specified by the user. Always returns imputed data by "Ensemble"
}
\description{
The function DreamAI_bagging imputes a dataset with missing values or NA's by bag imputaion with help of parallel processing. Pseudo datasets are generated having true missing (as in the original dataset) and pseudo missing and every such pseudo dataset is imputed by 7 different methods: KNN, MissForest, ADMIN, Brinn, SpectroFM, RegImpute and Ensemble (descriptions are included in the documentation of the function DreamAI).
}
\details{
This function can be run as parallel job in cluster. It generates and saves a .RData file containing the output from the current process in the location provided by the user, with the process number in the file name. If the user runs it in local computer multiple times, then changing the ProcessNumber everytime will generate and save .RData file with the given ProcessNumber.
}
\examples{
\dontrun{
data(datapnnl)
data<-datapnnl.rm.ref[1:100,1:21]
impute<- DreamAI_Bagging(data=data,k=10,maxiter_MF = 10, ntree = 100,maxnodes = NULL,maxiter_ADMIN=30,tol=10^(-2),gamma_ADMIN=NA,gamma=50,CV=FALSE,fillmethod="row_mean",maxiter_RegImpute=10,conv_nrmse = 1e-6,iter_SpectroFM=40,method=c("KNN","MissForest","ADMIN","Brinn","SpectroFM","RegImpute","Ensemble"),SamplesPerBatch=3,n.bag=2,save.out=TRUE,path="C:\\\\Users\\\\chowds14\\\\Desktop\\\\test_package\\\\",ProcessNum=1)
impute$Ensemble
}
}
