% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.r
\name{dummy}
\alias{dummy}
\title{Convert information as the right dummy variable numbers}
\usage{
dummy(PFERT = FALSE, BEDDED = FALSE, HERB = FALSE, BURN = FALSE,
  TYPEFERT = "NA", AGEFERT = NA, SOIL = "NA")
}
\arguments{
\item{PFERT}{TRUE if fertilized at planting. Default is FALSE.}

\item{BEDDED}{TRUE if bedded (site preparation). Default is FALSE.}

\item{HERB}{TRUE if herbicide was used. Default is FALSE.}

\item{BURN}{TRUE if burned was used. Default is FALSE.}

\item{TYPEFERT}{Type of fertilization at AGEFERT. N: only Nitrogen. NP: N and P. Default is NA.}

\item{AGEFERT}{Numeric value of stand age (in years) where fertilization is planned. Required if definied TYPEFERT.}

\item{SOIL}{Soil type. C: Group C. D: Group D. Default is NA.}
}
\value{
A list containing:
\itemize{
\item \code{Z1}Dummy variable if fertilized at planting then Z1 = 1.
\item \code{Z2}Dummy variable if bedded (site preparation) then Z2 = 1.
\item \code{Z3}Dummy variable if herbicide was used then Z3 = 1.
\item \code{ZB}Dummy variable if burned was used then ZB = 1.
\item \code{T1}Dummy variable if only Nitrogen fertilization at AGEFERT then T1 = 1.
\item \code{T2}Dummy variable if N and P fertilization at AGEFERT then T2 = 1.
\item \code{AGEFERT}Numeric value of stand age (in years) where fertilization is planned.
\item \code{S1}Dummy variable if soil is Goup C then S1 = 1.
\item \code{S2}Dummy variable if soil is Goup D then S2 = 1.
}
}
\description{
\code{dummy} Convert logical information or selectin to the right values as
categories that will be use in the G&Y Model.
}
\examples{
dummy(PFERT=TRUE, BEDDED=TRUE, HERB=FALSE, BURN=FALSE, TYPEFERT='NA', AGEFERT=NA, SOIL='C')
dummy(PFERT=FALSE, BEDDED=FALSE, HERB=FALSE, BURN=FALSE, TYPEFERT='NP', AGEFERT=15, SOIL='NA')
}
\references{
Gonzalez-Benecke et al. (2010) - Forest management effects on in situ and ex situ slash pine forest carbon balance
Forest Ecology and Management, 260(5), 795-805; https://doi.org/10.1016/j.foreco.2010.05.038
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
