% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm Implementation}
\usage{
euclidean(m, n)
}
\arguments{
<<<<<<< HEAD
\item{m}{number}

\item{n}{number}
}
\value{
Greatest Common Divisor of two Numbers by Euclidean Algorithm
}
\description{
Euclidean Algorithm Implementation to find the Greatest Common Divisor(GCD) of two numbers one and two param
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
