% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuroForecastHub-package.R
\docType{package}
\name{EuroForecastHub-package}
\alias{EuroForecastHub}
\alias{EuroForecastHub-package}
\title{EuroForecastHub: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})

Authors:
\itemize{
  \item European COVID-19 Forecast Hub contributors [copyright holder]
}

}
\keyword{internal}
