% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab4.R
\docType{package}
\name{fahaadUgurcan}
\alias{fahaadUgurcan}
\alias{fahaadUgurcan-package}
\title{A package for linear regression}
\description{
This package calculates regression coefficients by using given formula and data
}
\author{
Fahad Hameed and Ugurcan Lacin
}
