% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brinn.R
\name{gamma.CV}
\alias{gamma.CV}
\title{title: the selection process of parameter gamma using Cross Validation}
\usage{
\method{gamma}{CV}(x)
}
\description{
Arguments:
x: a matrix or data frame with either NAs or 0s as missings.
}
\details{
return: the best performed (with lowest MSE) gamma among
        (1e-3, 1e-2, 1e-1, 1, 10, 20, ..., 90, 100, 120, 140, ..., 480, 500)
}
