\name{get@PKGNAME@}
\alias{get@PKGNAME@}
\title{Accessor function - Greek mythology package: @TITLE@}
\description{Accessor function - Greek mythology package: @TITLE@}
\usage{
get@PKGNAME@(n)
}
\arguments{\item{n}{Integer.}}
\value{A character vector of the same size as \code{n}.}
\examples{
data(@PKGNAME@)
get@PKGNAME@(1:3)
}
\keyword{datasets}

