% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hub_config.R
\name{get_hub_config}
\alias{get_hub_config}
\title{Get the value of setting from config file}
\usage{
get_hub_config(setting, config_file = here::here("forecasthub.yml"))
}
\arguments{
\item{setting}{Setting for which to return the value. All values are returned
in a list if missing.}

\item{config_file}{Path to the config file}
}
\description{
Get the value of setting from config file
}
\examples{
# Get config for European COVID-19 forecast hub
get_hub_config(config_file = "https://raw.githubusercontent.com/epiforecasts/covid19-forecast-hub-europe/main/forecasthub.yml")

}
