% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/symbolics.R
\name{getSymbols}
\alias{getSymbols}
\title{Get symbols from a character}
\usage{
getSymbols(char, exclude = NULL)
}
\arguments{
\item{char}{Character vector (e.g. equation)}

\item{exclude}{Character vector, the symbols to be excluded from the return value}
}
\value{
character vector with the symbols
}
\description{
Get symbols from a character
}
\examples{
getSymbols(c("A*AB+B^2"))
}

