% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy knapsack}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{Dataframe with variables v and w.}

\item{W}{Integer as capacity.}
}
\value{
The Function Returns maximum knapsack value and corresponding elements that contributes to the value.
}
\description{
Calculates best values and weights for given capacity.
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Greedy_approximation_algorithm}
}
