% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_imputation_functions.R
\name{impute.ADMIN}
\alias{impute.ADMIN}
\title{Imputation of Missing Protein Abundances using ADMIN (Abundance  Dependent  Missing Imputation Mechanism)}
\usage{
impute.ADMIN(data, data.ini = NA, gamma = NA, k, maxiter_ADMIN, tol)
}
\arguments{
\item{data}{dataset in the form of a matrix or data frame with either NAs or 0s as missings}

\item{data.ini}{initial dataset, set to be NA as default}

\item{gamma}{parameter of the non-ignorable missing mechanism}

\item{k}{number of neighbors to be used in the imputation (default=10)}

\item{maxiter_ADMIN}{maximum number of iteration to be performed if the stopping criteria is not met beforehand}

\item{tol}{convergence threshold}
}
\value{
the imputed version of the dataset
}
\description{
The function impute.ADMIN imputes a dataset with missing values or NA's using ADMIN
}
\examples{
\dontrun{
data(datapnnl)
data<-datapnnl.rm.ref[1:100,1:21]
impute.ADMIN(data = as.matrix(data),k = 10, maxiter_ADMIN = 30, tol = 10^(-2))
}
}
