% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brinn.R
\name{impute.Brinn}
\alias{impute.Brinn}
\title{Imputation of Missing Protein Abundances using IRNN-SCAD Algorithm}
\usage{
impute.Brinn(x, gamma = gamma, CV = CV)
}
\arguments{
\item{x}{dataset in the form of a matrix or data frame with either NAs or 0s as missings}

\item{gamma}{the parameter of  the nonconvex surrogate function of L0 norm.
The default \code{gamma = 50} works well for a 7000 * 80 protein-sample matrix.
Can be automatically fit by setting \code{CV = TRUE}.}

\item{CV}{a logical value indicating whether to fit the best gamma with Cross Validation.
The process can be time consuming.}
}
\value{
the imputed version of the dataset
}
\description{
The function impute.Brinn imputes a dataset with missing values or NA's using IRNN-SCAD Algorithm
}
\examples{
\dontrun{
data(datapnnl)
data<-datapnnl.rm.ref[1:100,1:21]
impute.Brinn(data=as.matrix(data), gamma = 50, CV = FALSE)
}
}
