% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegImpute.R
\name{impute.RegImpute}
\alias{impute.RegImpute}
\title{Imputation of Missing Protein Abundance using Glmnet Ridge Regression.}
\usage{
impute.RegImpute(data, fillmethod, maxiter_RegImpute, conv_nrmse)
}
\arguments{
\item{data}{dataset in the form of a matrix or data frame with either NAs or 0s as missings.}

\item{fillmethod}{a string identifying the method to be used that could be "row_mean" or "zeros", with "row_mean" being the default. It throws an warning if "row_median" is used.}

\item{maxiter_RegImpute}{a integer identifying maximum number of iterations to reach convergence}
}
\value{
the imputed version of the dataset
}
\description{
The function impute.RegImpute imputes a dataset with missing values or NA's using Glmnet Ridge Regression
}
\examples{
\dontrun{
data(datapnnl)
data<-datapnnl.rm.ref[1:100,1:21]
impute.RegImpute(data=as.matrix(data), fillmethod = "row_mean", maxiter_RegImpute = 10,conv_nrmse = 1e-06)
}
}
