% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{jacobianSymb}
\alias{jacobianSymb}
\title{Compute Jacobian of a function symbolically}
\usage{
jacobianSymb(f, variables = NULL)
}
\arguments{
\item{f}{named vector of type character, the functions}

\item{variables}{other variables, e.g. paramters, f depends on. If variables is
given, f is derived with respect to variables instead of \code{names(f)}}
}
\value{
named vector of type character with the symbolic derivatives
}
\description{
Compute Jacobian of a function symbolically
}
\examples{
jacobianSymb(c(A="A*B", B="A+B"))
jacobianSymb(c(x="A*B", y="A+B"), c("A", "B"))
}

