% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic Knapsack Algorithm}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{Dataframe with variables v and w.}

\item{W}{Integer as capacity.}
}
\value{
Returns a list with best value and combination.
}
\description{
Calculates best values and weights for given capacity.
}
