% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear Regression}
\description{
You can have Reference Class containing some calculations by giving formula and data.
}
\section{Fields}{

\describe{
\item{\code{formula}}{Formula}

\item{\code{data}}{A data frame}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{This function returns regression coefficients}

\item{\code{plot()}}{This function plots two graphs, such as Fitted values vs Residuals and Scale Location by using given formula and data in initialization.}

\item{\code{pred()}}{This function returns fitted value.}

\item{\code{print()}}{This function prints regression coefficients by using given formula and data in initialization.}

\item{\code{resid()}}{This function returns residuals value.}

\item{\code{summary()}}{This function prints the coefficients with their standard error, t-value and p-value.}
}}

\examples{
data(iris)
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$print()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$pred()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$summary()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$resid()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$coef()
linreg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris)$plot()
}
