% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birnn.R
\name{mc.IRNN}
\alias{mc.IRNN}
\title{title: IRNN algorithm for matrix completion
Arguments:
fun_sg: {scad_sg, mcp_sg}, the supergradient of the selected nonconvex surrogate function of L0 norm
y: a vector with all entry values
IDX: a vector with all entry positions
m: the number of rows present in the incomplete dataset (proteins)
n: the number of columns present in the incomplete dataset (samples)
gamma: 'fun_sg''s parameter
x_initial: the initial matrix}
\usage{
mc.IRNN(fun_sg, y, IDX, m, n, gamma, x_initial)
}
\description{
return: an imputed matrix
}
