% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.BA.r
\name{module.BA}
\alias{module.BA}
\title{Predicts and projects stand basal area based on stand-level information.}
\usage{
module.BA(N0 = NA, HDOM0 = NA, AGE0 = NA, Z1 = 0, Z2 = 0, Z3 = 0,
  ZB = 0, T1 = 0, T2 = 0, AGEFERT = NA, S1 = 0, S2 = 0,
  projection = FALSE, BA0 = NA, N1 = NA, HDOM1 = NA)
}
\arguments{
\item{N0}{Numeric value of number of trees per acre at age 0 (or initial age).}

\item{HDOM0}{Numeric value of Dominant Height (feet) at age 0.}

\item{AGE0}{Numeric value of initial stand age (years).}

\item{Z1}{If fertilized at planting then Z1 = 1. Default is 0.}

\item{Z2}{If bedded (site preparation) then Z2 = 1. Default is 0.}

\item{Z3}{If herbicide was used then Z3 = 1. Default is 0.}

\item{ZB}{If burned was used then ZB = 1. Default is 0.}

\item{T1}{If only Nitrogen fertilization at AGEFERT then T1 = 1. Default is 0.}

\item{T2}{If N and P fertilization at AGEFERT then T2 = 1. Default is 0.}

\item{AGEFERT}{Numeric value of stand age (in years) where fertilization is planned. Required if T1=1 or T2=1.}

\item{S1}{If soil is Goup C then S1 = 1. Default is 0.}

\item{S2}{If soil is Goup D then S2 = 1. Default is 0.}

\item{projection}{If TRUE then model projection from provided BA0 is executed for a 1 year increment. Default: FALSE.}

\item{BA0}{Numeric value of Basal Area (ft2/acre) at age 0 (required for model projection).}

\item{N1}{Numeric value of number of trees per acre at age 1 for projection.}

\item{HDOM1}{Numeric value of Dominant Height (ft) at age 1 for projection.}
}
\value{
A list containing:
\itemize{
\item \code{BA0} Predicted Basal Area at age 0 (ft2/acre).
\item \code{BA1} Projected Basal Area at age 1 (ft2/acre).
}
}
\description{
\code{module.BA} Predicts and projects stand basal area based on number of trees per acre and
dominant height at the stand-level considering silvicultural information.
}
\examples{
# Example 1 - Predicting BA
module.BA(N0=493, HDOM0=41.57, AGE0=12,Z1=1, Z2=1, Z3=1, ZB=1, T1=1, T2=0,
AGEFERT=1, S1=0, S2=0)$BA0

# Example 2 - Projecting BA
SI <- stand.SITE(HDOM=41.57, AGE=12)$SI
HDOM1 <-stand.SITE(SI=SI, AGE=13)$HDOM
N1 <- module.N(N0=493, AGE0=12, AGE1=13)$N1
module.BA(N0=493, HDOM0=41.57, AGE0=12, projection=TRUE, BA0=99.74, N1=N1, HDOM1=HDOM1)$BA1
}
\references{
Gonzalez-Benecke et al. (2010) - Forest management effects on in situ and ex situ slash pine forest carbon balance
Forest Ecology and Management, 260(5), 795-805; https://doi.org/10.1016/j.foreco.2010.05.038
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
