% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.N.r
\name{module.N}
\alias{module.N}
\title{Predicts the number of trees (mortality) at next age based on stand-level informtion.}
\usage{
module.N(N0 = NA, AGE0 = NA, AGE1 = NA, THINNING = FALSE, AGET = NA)
}
\arguments{
\item{N0}{Numeric value of number of trees per hectare at age 0.}

\item{AGE0}{Numeric value of intial stand age or age 0 (years).}

\item{AGE1}{Numeric value of final stand age or age 1 for prediction (years).}

\item{THINNING}{If TRUE then thinning happened. Default is FALSE.}

\item{AGET}{Optional numeric value of stand age (in years) where thinning is planned.}
}
\value{
A value with the number of trees per hectare at age 1 (N1).
}
\description{
\code{module.N} Estimates the number of trees (mortality) from initial age (AGE0) to next age (AGE1) based on stand level
information based on the equation:
N1 = N0 x  exp(-(-c1 -c2 x Z) x ((AGE1^c3)-(AGE0^c3)))
}
\examples{
module.N(N0=376, AGE0=12, AGE1=21)$N1  # Whitout thinning
module.N(N0=376, AGE0=12, AGE1=31, THINNING=TRUE, AGET=18)$N1  # Whitout thinning
}
\references{
Gonzalez-Benecke et al. (2010) - Forest management effects on in situ and ex situ slash pine forest carbon balance
Forest Ecology and Management, 260(5), 795-805; https://doi.org/10.1016/j.foreco.2010.05.038
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
