% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.simulation.r
\name{module.simulation}
\alias{module.simulation}
\title{Module that performs the simulations by projections at the desired age starting from initial conditions.}
\usage{
module.simulation(stand = NULL)
}
\arguments{
\item{stand}{List with information originated from \code{\link{module.input}} containing:
SI, AGE0, HDOM0, BA0, N0, QD0, SDIR0, VOL_0B0, VOL_IB0, VOLm_0B0, VOLm_IB0,
AGEF, THINNING, AGET, BAR, t, d, method.}
}
\value{
A list containing a data frame (sim.stand) containing the columns below with simulations from
AGE0 to AGEF and a logical value indicating if thinning was requested.
\itemize{
\item \code{AGE}     Stand age form simulations from AGE0 to AGEF (years).
\item \code{HDOM}    Dominant Height (ft).
\item \code{SI}      Site Index (ft).
\item \code{BA}      Basal Area (ft2/ha).
\item \code{N}       Number of trees per hectare.
\item \code{QD}      Mean quadratic diameter (in).
\item \code{SDIR}    Relative stand density index (\%).
\item \code{VOL_OB}  Total stand-level volume outside bark (ft3/ha).
\item \code{VOL_IB}  Total stand-level volume inside bark (ft3/ha).
\item \code{VOLm_OB} Merchantable stand-level volume outside bark (ft3/ha).
\item \code{VOLm_IB} Merchantable stand-level volume inside bark (ft3/ha).
\item \code{CI}      Competition Index (required for THINNING=TRUE).
\item \code{BAU}     Basal Area (ft2/ha) of unthinned counterpart (required for THINNING=TRUE).
\item \code{NU}      Number of trees per hectare of unthinned counterpart (required for THINNING=TRUE).
}
}
\description{
\code{module.simulation} Performs model projections of a stand to a final age based on information provided
by the input module that contains:  SI, AGE0, HDOM0, BA0, N0, QD0, SDIR0, VOL_0B0, VOL_IB0,
VOLm_0B0, VOLm_IB0, AGEF, THINNING, AGET, BAR, t, d and method.
Provides with a data frame with all relevant information for simulations from AGE0 to AGEF and thinning (if requested).
}
\examples{
# Example 1 - Input from plot-level data (not thinning)
sim1 <- module.input(TYPE='PLOT', BA0=93, HDOM0=54, AGE0=18, N0=282, AGEF=19)
module.simulation(stand=sim1)

# Example 2 - Input from tree-level data
treedata <- subset(treedata, is.na(DBH)==FALSE)
sim2 <- module.input(TYPE='TREE', TREEDATA=treedata, AREA=3240, AGE0=5, AGEF=21)
module.simulation(stand=sim2)

# Example 3 - Input from plot-level data (with thinning)
sim3 <- module.input(TYPE='PLOT', BA0=93, SI=70, AGE0=18, N0=282, AGEF=25,
                     THINNING=TRUE, AGET=22, NT=70)
sims <- module.simulation(stand=sim3)$sim.stand
sims
plot(sims$AGE,sims$BAU,type="l",xlim=c(15,30),ylim=c(0,10000),col=2,
     xlab='Age (years)', ylab='Basal Area (feet2/acre)')
par(new=TRUE)
plot(sims$AGE,sims$BA,type="l",xlim=c(15,30),ylim=c(0,10000),col=1,
     xlab='Age (years)', ylab='Basal Area (feet2/acre)')
par(new=FALSE)
}
\references{
Gonzalez-Benecke et al. (2012) - Modeling Survival, Yield, Volume Partitioning and Their Response to Thinning for Longleaf Pine Plantations.
Forests, 3(4), 1104-1132; doi:10.3390/f3041104
}
\seealso{
\code{\link{module.input}}
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
