% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.VOL.r
\name{module.VOL}
\alias{module.VOL}
\title{Calculates the total stand-level volume based on stand-level information.}
\usage{
module.VOL(HDOM = NA, BA = NA, N = NA, AGE = NA, methodVOL = 1)
}
\arguments{
\item{HDOM}{Numeric value of Site Index (ft) (Dominant Height of the plot at age 25 years).}

\item{BA}{Numeric value of Basal Area of the plot (ft2/acre).}

\item{N}{Numeric value of number of trees per acre.}

\item{AGE}{Numeric value of Age of the plot (in years).}

\item{methodVOL}{Numeric value that identifies the method to compute the total stand-level.\itemize{
\item 1: parametrized DBH-height model that requires HDOM, N, BA and AGE.
\item 2: parametrized DBH-height model that requires only HDOM and BA.
}
Default method = 1.}
}
\value{
A list containing the parameters:
\itemize{
 \item \code{VOL_OB} Total stand-level volume outside bark (ft3/acre).
 \item \code{VOL_IB} Total stand-level volume inside bark (ft3/acre).
}
}
\description{
\code{module.VOL} Calculates the total stand-level volume inside and outside bark for
a Slash pine stands based a selected method.
}
\examples{
# Example 1 - Method 1
module.VOL(HDOM=46, BA=100, N=376, AGE=12, methodVOL=1)

# Example 2 - Method 2
module.VOL(HDOM=46, BA=100, N=376, AGE=12, methodVOL=2)

# Example 3 - Method 3
module.VOL(HDOM=46, BA=100, N=376, AGE=12, methodVOL=3)
}
\references{
Gonzalez-Benecke et al. (2010) - Forest management effects on in situ and ex situ slash pine forest carbon balance
Forest Ecology and Management, 260(5), 795-805; https://doi.org/10.1016/j.foreco.2010.05.038
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
