% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.VOLm.r
\name{module.VOLm}
\alias{module.VOLm}
\title{Calculates the stand-level merchantable volume based on stand-level information.}
\usage{
module.VOLm(N = NA, QD = NA, t = NA, d = NA, VOL_OB = NA,
  VOL_IB = NA)
}
\arguments{
\item{N}{Numeric value of number of trees per acre.}

\item{QD}{Numeric value of mean Quadratic Diameter (in).}

\item{t}{Numeric value top stem diameter outside bark for merchantability limit (in).}

\item{d}{Numeric value of a DBH threshold limit for merchantable trees (in).}

\item{VOL_OB}{Total stand-level volume outside bark (ft3/acre).}

\item{VOL_IB}{Total stand-level volume inside bark (ft3/acre).}
}
\value{
A list containing the parameters:
\itemize{
\item \code{VOLm_OB} Merchantable stand-level volume outside bark (ft3/acre).
\item \code{VOLm_IB} Merchantable stand-level volume inside bark (ft3/acre).
}
}
\description{
\code{module.VOLm} Calculates the stand-level merchantable volume inside and outside bark from a plot
based on the equation:  VOLm = VOL x exp(m1 x ((t/QD)^m2) + m3 x (N^m4) x (d/QD)^m5)
}
\examples{
VOL_OB<-module.VOL(HDOM=46, BA=100, N=376, AGE=12)$VOL_OB
VOL_IB<-module.VOL(HDOM=46, BA=100, N=376, AGE=12)$VOL_IB
module.VOLm(N=376, QD=7, t=1, d=1.96, VOL_OB=VOL_OB, VOL_IB=VOL_IB)
}
\references{
Gonzalez-Benecke et al. (2010) - Forest management effects on in situ and ex situ slash pine forest carbon balance
Forest Ecology and Management, 260(5), 795-805; https://doi.org/10.1016/j.foreco.2010.05.038
}
\seealso{
\code{\link{module.VOL}}
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
