% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.tree.r
\name{prepare.tree}
\alias{prepare.tree}
\title{Checks and prepares tree-level data from a single plot and calculates some stand-level parameters.}
\usage{
prepare.tree(TREEID = NA, DBH = NA, HT = NA, AREA = NA, AGE = NA,
  methodHT = 2)
}
\arguments{
\item{TREEID}{Vector of unique tree identification.}

\item{DBH}{Vector of diameter at breast height (DBH, in). Must be complete and have the same size and order as TREEID.}

\item{HT}{Vector of total height (ft). Must be of the same size and order as TREEID.}

\item{AREA}{Numeric value of area of the inventory plot (ft2).}

\item{AGE}{Numeric value of stand age (years). Required if method = 1.}

\item{methodHT}{Numeric value that identifies the method to estimate missing heights. 1: parametrized DBH-height model,
2: fits a simple DBH-height model from available measurements. Default method = 2.}
}
\value{
A list containing the following:
\itemize{
\item \code{BA}   Basal Area (ft2/acre).
\item \code{N}    Number of trees per acre.
\item \code{HDOM} Dominant Height (ft).
\item \code{tree.table} Data frame with all tree data an observed heights (for the ones provided) and estimated heights
(for those missing). The data frame contains the columns: TREEID, DBH and HT.
}
}
\description{
\code{prepare.tree} Checks and prepares tree-level data from a single plot and then calculates stand-level
parameters such as basal area, number of trees per hectarea and dominant height. The provided vector of total heights
can have missing information. If there are missing trees, there are two methods to use: 1) Estimates heights according
to a parametrized DBH-height model, or 2) Estimates heights by fitting a simple DBH-height model that requires at least
10 measurements. Missing values are indentified as 'NA'.
}
\examples{
# Example - Stand-level information from inventory data
treedata <- subset(treedata, is.na(DBH)==FALSE)
TREEID <- treedata$TREEID
DBH <- treedata$DBH
HT <- treedata$HT
prepare.tree(TREEID=TREEID, DBH=DBH, HT=HT, AREA=3240, AGE=5, methodHT=2)
}
\seealso{
\code{\link{tree.HT}}
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
