% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{prodSymb}
\alias{prodSymb}
\title{Compute matrix product symbolically}
\usage{
prodSymb(M, N)
}
\arguments{
\item{M}{matrix of type character}

\item{N}{matrix of type character}
}
\value{
Matrix of type character, the matrix product of M and N
}
\description{
Compute matrix product symbolically
}

