\name{ProMap-package}
\alias{ProMAP-package}
\alias{ProMAP}
\docType{package}
\title{  A Regularized Linear Mixed Effect Model (Mixed remMap) 
}
\description{
This package estimates a regularized linear mixed effect model for hub networks.
}
\details{
\tabular{ll}{
Package: \tab ProMAP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-25\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Weiping Ma, Sung Won Han, Hua Zhong, and Pei Wang

Maintainer: Pei Wang <pei.wang@mssm.edu>
}
\references{
Ma, W., Chen, L.S., Han, S.W., Lin, C., Yan, P., Paulovich, A., CPTAC consortium, Zhong, H., and Wang, P. (2015) \emph{A Penalized Multivariate Linear Mixed Effects Model for Integrative Proteo-genomics Analysis Based on iTRAQ Data}
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
