% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.R
\name{reduceSensitivities}
\alias{reduceSensitivities}
\title{reduceSensitivities}
\usage{
reduceSensitivities(sens, vanishing)
}
\arguments{
\item{sens}{Named character, the sensitivity equations}

\item{vanishing}{Character, names of the vanishing sensitivities}
}
\value{
Named character, the sensitivity equations with zero entries for vanishing
sensitivities.
}
\description{
reduceSensitivities
}
\details{
Given the set \code{vanishing} of vanishing sensitivities, the algorithm
determins sensitivities that vanish as a consequence of the first set.
}

