% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{replaceNumbers}
\alias{replaceNumbers}
\title{Replace integer number in a character vector by other double}
\usage{
replaceNumbers(x)
}
\arguments{
\item{x}{vector of type character, the object where the replacement should take place}
}
\value{
vector of type character, conserves the names of x.
}
\description{
Replace integer number in a character vector by other double
}
