% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{replaceOperation}
\alias{replaceOperation}
\title{Replace a binary operator in a string by a function}
\usage{
replaceOperation(what, by, x)
}
\arguments{
\item{what}{character, the operator symbol, e.g. "^"}

\item{by}{character, the function string, e.g. "pow"}

\item{x}{vector of type character, the object where the replacement should take place}
}
\value{
vector of type character
}
\description{
Replace a binary operator in a string by a function
}
\examples{
replaceOperation("^", "pow", "(x^2 + y^2)^.5")
}
