% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{replaceSymbols}
\alias{replaceSymbols}
\title{Replace symbols in a character vector by other symbols}
\usage{
replaceSymbols(what, by, x)
}
\arguments{
\item{what}{vector of type character, the symbols to be replaced, e.g. c("A", "B")}

\item{by}{vector of type character, the replacement, e.g. c("x[0]", "x[1]")}

\item{x}{vector of type character, the object where the replacement should take place}
}
\value{
vector of type character, conserves the names of x.
}
\description{
Replace symbols in a character vector by other symbols
}
\examples{
replaceSymbols(c("A", "B"), c("x[0]", "x[1]"), c("A*B", "A+B+C"))
}
