% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegImpute.R
\name{returnPresentIndices}
\alias{returnPresentIndices}
\title{Return current column indices where values are not missing.}
\usage{
returnPresentIndices(col, data, filled)
}
\arguments{
\item{col}{A number indicating current columns index.}

\item{data}{A matrix, The original data matrix with NaNs.}

\item{filled}{A matrix, The data matrix after running backfill function.}
}
\value{
A list of indices corresponding to current working column values.
}
\description{
Return current column indices where values are not missing.
}
