% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegImpute.R
\name{returnTrainingSet}
\alias{returnTrainingSet}
\title{Return training set for current column.}
\usage{
returnTrainingSet(col, filled, present_indices)
}
\arguments{
\item{col}{A number indicating current columns index.}

\item{filled}{A matrix, The data matrix after running backfill function.}

\item{present_indices}{A list containing indices in current column where values are not missing.}
}
\value{
A matrix.
}
\description{
Return training set for current column.
}
