% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Linear Regression}
\description{
You can have Reference Class containing some calculations by giving formula and data.
}
\section{Fields}{

\describe{
\item{\code{formula}}{Formula}

\item{\code{data}}{A data frame}
}}

\section{Methods}{

\describe{
\item{\code{print()}}{This function prints regression coefficients by using given formula and data in initialization.}
}}

