% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ensemble.R
\name{run_ensemble}
\alias{run_ensemble}
\title{Run ensembling methods}
\usage{
run_ensemble(
  method = "mean",
  forecast_date,
  exclude_models = NULL,
  return_criteria = TRUE,
  verbose = FALSE,
  exclude_designated_other = TRUE,
  ...
)
}
\arguments{
\item{method}{name of ensembling method}

\item{forecast_date}{date or character}

\item{exclude_models}{optional character vector to exclude over all dates,
or data.frame with cols model and forecast_date, to exclude for specific
dates}

\item{return_criteria}{logical : whether to return a model/inclusion criteria
grid as well as the ensemble forecast (default \code{TRUE})}

\item{verbose}{Logical determining whether diagnostic messages should
be printed while running (defaults to \code{FALSE}).}

\item{exclude_designated_other}{logical: whether to exclude models designated
as "other" in their metadata file (default \code{TRUE})}

\item{...}{arguments passed to \code{\link[=create_ensemble_relative_skill]{create_ensemble_relative_skill()}}}
}
\description{
Run ensembling methods
}
\details{
Used to create a single ensemble forecast.
Takes a forecast date and loads all forecasts for the preceding week, then:
Filters models according to criteria
Ensembles forecasts according to given method
Formats ensemble forecast
Returns ensemble forecast and optionally the criteria for inclusion
}
