% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiple_ensembles.R
\name{run_multiple_ensembles}
\alias{run_multiple_ensembles}
\title{Runs and saves past ensembles for specified methods/dates.}
\usage{
run_multiple_ensembles(forecast_dates, methods, ...)
}
\arguments{
\item{forecast_dates}{Dates vector}

\item{methods}{character vector of method/s supported in \code{\link[=run_ensemble]{run_ensemble()}}
(as in the named folders of code/ensemble/forecasts)}

\item{...}{arguments passed to \code{\link[=run_ensemble]{run_ensemble()}}}
}
\value{
a list of ensembles with forecasts, method, forecast date, criteria
}
\description{
Wrapper around \code{\link[=run_ensemble]{run_ensemble()}} that allows multiple methods/forecast dates.
}
