% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.SDI.r
\name{stand.SDI}
\alias{stand.SDI}
\title{Calculates relative stand density index for Slash pine (\emph{Pinus ellioti})}
\usage{
stand.SDI(N = NA, QD = NA)
}
\arguments{
\item{N}{Numeric value of the number of trees per acre.}

\item{QD}{Numeric value of mean quadratic diameter (in).}
}
\value{
A value for the relative stand density index (RSDI, \%).
}
\description{
\code{stand.SDI} Calculates the relative stand density index (RSDI, \%) for Slash in a single plot using the
expression: RSDI = 100xNx(QD/10)^1.605/SDImax. For Slash pine (\emph{Pinus ellioti}) SDImax is 400 trees/acre.
All parameters are required to compute the RSDI.
}
\examples{
stand.SDI(QD=7, N=376)
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
