% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.SITE.r
\name{stand.SITE}
\alias{stand.SITE}
\title{Calculates dominant height for a plot.}
\usage{
stand.SITE(HDOM = NA, SI = NA, AGE = NA, Z1 = 0, Z2 = 0, Z3 = 0,
  T1 = 0, T2 = 0, AGEFERT = NA)
}
\arguments{
\item{HDOM}{Numeric value of mean Dominant Height (ft).}

\item{SI}{Numeric value of Site Index (ft) (Dominant Height of the plot at age 25 years).}

\item{AGE}{Numeric value of stand Age (years).}

\item{Z1}{If fertilized at planting then Z1 = 1. Default is 0.}

\item{Z2}{If bedded (site preparation) then Z2 = 1. Default is 0.}

\item{Z3}{If herbicide was used then Z3 = 1. Default is 0.}

\item{T1}{If only Nitrogen fertilization at AGEFERT then T1 = 1. Default is 0.}

\item{T2}{If N and P fertilization at AGEFERT then T2 = 1. Default is 0.}

\item{AGEFERT}{Numeric value of stand age (in years) where fertilization is planned. Required if T1=1 or T2=1.}
}
\value{
A list containing the parameters:
\itemize{
\item \code{HDOM} Dominant Height (ft).
\item \code{SI} Site Index (ft).
}
}
\description{
\code{stand.SITE} Calculates dominant height, based on site index (at reference age 25 year) and
stand age for a plot considering silvicultural managements (i.e. site preparation, fertilization).
}
\examples{
# Example 1 - No fertilization
stand.SITE(SI=70,AGE=18)$HDOM

# Example 2 - With fertilization
stand.SITE(SI=70,AGE=18,T2=1, AGEFERT=15)$HDOM

# Example 3 - Site Index
stand.SITE(HDOM=54,AGE=18)$SI
}
\references{
Gonzalez-Benecke et al. (2010) - Forest management effects on in situ and ex situ slash pine forest carbon balance
Forest Ecology and Management, 260(5), 795-805; https://doi.org/10.1016/j.foreco.2010.05.038
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
