% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.STAND.r
\name{stand.STAND}
\alias{stand.STAND}
\title{Calculates the mean quadratic diameter, basal area or number of trees.}
\usage{
stand.STAND(BA = NA, N = NA, QD = NA)
}
\arguments{
\item{BA}{Numeric value of basal area of the plot (ft2/acre).}

\item{N}{Numeric value of number of trees per acre.}

\item{QD}{Numeric value of mean quadratic diameter (in).}
}
\value{
A list containing the parameters:
\itemize{
\item \code{BA} Basal Area (ft2/acre).
\item \code{N}  Number of trees per acre.
\item \code{QD} Mean Quadratic Diameter (in).
}
}
\description{
\code{stand.STAND} Calculates the mean quadratic diameter, basal area or number of trees based on stand-level information
of the plot. At least two parameters are required to calculate the third missing parameter.
}
\examples{
stand.STAND(QD=7, N=376)$BA     # Missing BA
stand.STAND(BA=100, QD=7)$N     # Missing N
stand.STAND(BA=100, N=376)$QD   # Missing QD
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
