% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolics.R
\name{sumSymb}
\alias{sumSymb}
\title{Compute matrix sumSymbolically}
\usage{
sumSymb(M, N)
}
\arguments{
\item{M}{matrix of type character}

\item{N}{matrix of type character}
}
\value{
Matrix of type character, the matrix sum of M and N
}
\description{
Compute matrix sumSymbolically
}
