% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cOde.R
\name{sundialsIncludes}
\alias{sundialsIncludes}
\title{Infos and includes for sundials::cvodes C++ code: ODE, Jacobian.}
\usage{
sundialsIncludes()
}
\value{
C++ source code as a character vector.
}
\description{
Infos and includes for sundials::cvodes C++ code: ODE, Jacobian.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

