% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cOde.R
\name{sundialsSensOde}
\alias{sundialsSensOde}
\title{Implement sensitivities for an ODE system, sundials::cvodes.}
\usage{
sundialsSensOde(f, variablesOde, variablesSens, parameters)
}
\arguments{
\item{f}{Named character vector containing the right-hand sides of the ODE. 
You may use the key word.}

\item{variablesOde}{Variables appearing on the ODE of the dynamic system.}

\item{variablesSens}{Variables appearing on the ODE of the sensitivities of
the dynamic system.}

\item{parameters}{Parameters appearing in the ODE.}
}
\value{
C++ source code as a character vector.
}
\description{
Implement sensitivities for an ODE system, sundials::cvodes.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}
