% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.HDOM.r
\name{tree.HDOM}
\alias{tree.HDOM}
\title{Calculate dominant height for a plot based on tree-level data.}
\usage{
tree.HDOM(HT, DBH, AREA)
}
\arguments{
\item{HT}{Vector of tree heights (ft) for a given plot (must be complete).}

\item{DBH}{Vector of diameter at breast height (DBH, in). Must be complete and have the same size and order of HT.}

\item{AREA}{Numeric value of area of the inventory plot (ft2).}
}
\value{
A value with Dominant Height (HDOM, ft) for the plot.
}
\description{
\code{tree.HDOM} Calculate the dominant height for a plot based on tree-level data based on the top 25th percentile
of the tree heights. The provided vector of heights should be complete without missing data.
}
\examples{
treedata <- subset(treedata, is.na(DBH)==FALSE)
HT <- treedata$HT
DBH <- treedata$DBH
AREA <- 3240
tree.HDOM(HT=HT, DBH=DBH, AREA=AREA)
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
