% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.HDOM25.r
\name{tree.HDOM25}
\alias{tree.HDOM25}
\title{Calculate dominant height for a plot based on tree-level data.}
\usage{
tree.HDOM25(HT)
}
\arguments{
\item{HT}{Vector of tree heights (ft) for a given plot (must be complete).}
}
\value{
A value with Dominant Height (HDOM, ft) for the plot.
}
\description{
\code{tree.HDOM25} Calculate the dominant height for a plot based on tree-level data based on the top 25th percentile
of the tree heights. The provided vector of heights should be complete without missing data.
}
\examples{
HT <- treedata$HT[!is.na(treedata$HT)]
tree.HDOM25(HT=HT)
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
