% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.HT.r
\name{tree.HT}
\alias{tree.HT}
\title{Estimates the total height of trees for a single plot.}
\usage{
tree.HT(DBH, HT, AREA = NA, AGE = NA, BA = NA, N = NA, methodHT = 2)
}
\arguments{
\item{DBH}{Vector of diameter at breast height (DBH, in). Must be complete and have the same size and order as TREEID.}

\item{HT}{Vector of total height (ft). Must be of the same size and order as TREEID.}

\item{AREA}{Numeric value of area of the inventory plot (ft2).}

\item{AGE}{Numeric value of stand age (years). Required if methodHT = 1.}

\item{BA}{Numeric value of Basal Area (ft2/acre). Required if methodHT = 1.}

\item{N}{Numeric value of number of trees per acre. Required if methodHT = 1.}

\item{methodHT}{Numeric value that identifies the method to estimate missing heights. 1: parametrized DBH-height model
that requires DBH, BA and AGE, 2: fits a simple DBH-height model from available measurements using the equation:
ln(HT) = b0 + b1/DBH. Default methodHT = 2.}
}
\value{
A list containing the following:
\itemize{
\item \code{HTFIN} A vector of final tree heigths (ft), replacing the missing values for estimated heigths and retaining the observed heights.
\item \code{r2} A value with the coefficient of determination from the fitting the DBH-height model when methodHT = 2.
}
}
\description{
\code{tree.HT} Estimates the total height of trees that have missing height from tree-level data.
For the missing trees there are two methods to use: 1) Estimates heights according
to a parametrized DBH-height model, or 2) Estimates heights by fitting a simple DBH-height model that requires at least
10 measurements. Missing values are indentified as 'NA'.
}
\examples{
# Example 1 - Method 1 - Parametrized DBH-height model
DBH <- c(9.3,11.1,15.5,9,14.8,27.3,11.4,6.6,12.6,17.5,6.3,7.2,11.5,13.6,7.3,12,11.9,8.1,7.6,5)
HT <- c(11.8,12.3,NA,NA,15.3,18,12,NA,14.5,NA,NA,NA,NA,NA,10.3,14.6,NA,NA,NA,NA)
tree.HT(DBH=DBH, HT=HT, AREA=200, AGE=47, methodHT=1)

# Example 2 - Method 2 - Simple DBH-height model
DBH <- c(9.3,11.1,15.5,9,14.8,27.3,11.4,6.6,12.6,17.5,6.3,7.2,11.5,13.6,7.3,12,11.9,8.1,7.6,5)
HT <- c(11.8,12.3,NA,NA,15.3,18,12,NA,14.5,NA,NA,NA,NA,NA,10.3,14.6,NA,NA,NA,NA)
tree.HT(DBH=DBH, HT=HT, methodHT=2)
}
\references{
Gonzalez-Benecke et al. (2014) - IParameterization of the 3-PG model for Pinus elliottii stands using alternative methods to estimate fertility rating, biomass
partitioning and canopy closure. Forest Ecology and Management, 327(1): 55–75; https://doi.org/10.1016/j.foreco.2014.04.030
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
