\name{UgurcanFahaadLab3-package}
\alias{UgurcanFahaadLab3-package}
\alias{UgurcanFahaadLab3}
\docType{package}
\title{
\packageTitle{UgurcanFahaadLab3}
}
\description{
\packageDescription{UgurcanFahaadLab3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{UgurcanFahaadLab3}
\packageIndices{UgurcanFahaadLab3}
  There are two functions, such as, euclidean and dijkstra.
}
\author{
\packageAuthor{UgurcanFahaadLab3}

Maintainer: \packageMaintainer{UgurcanFahaadLab3}
}
\references{
  https://en.wikipedia.org/wiki/Euclidean_algorithm
  https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm
}

\keyword{ package }
\seealso{


}
\examples{

}
