% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ensemble_criteria.R
\name{use_ensemble_criteria}
\alias{use_ensemble_criteria}
\title{Load models according to inclusion criteria before ensembling}
\usage{
use_ensemble_criteria(
  forecasts,
  exclude_models = NULL,
  exclude_designated_other = TRUE,
  return_criteria = TRUE
)
}
\value{
a list:
\itemize{
\item "forecasts" = tibble, all forecasts passing inclusion criteria
\item "criteria" = tibble, model names and criteria assessment
}
}
\description{
Load models according to inclusion criteria before ensembling
}
\details{
Steps:
Currently, models included based on having:
\enumerate{
\item All quantiles
\item 4 horizons
\item Not manually specified for exclusion
\item Not the hub ensemble
}
}
