% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Plot mean delay}
\usage{
visualize_airport_delays()
}
\description{
visualizes the mean delay of flights for different airports by longitude and latitude.
}
\section{Fields}{

\describe{
\item{\code{no}}{argument}
}}

\examples{
visualize_airport_delays()
}
