% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weighted_average}
\alias{weighted_average}
\title{Compute weighted mean/median}
\usage{
weighted_average(..., average = c("mean", "median"))
}
\arguments{
\item{...}{values to average}

\item{average}{Method to average. One of \code{"mean"} or \code{"median"}}
}
\description{
Compute weighted mean/median
}
