% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\format{An object of class \code{data.frame} with 18 rows and 3 columns.}
\usage{
wiki_graph
}
\arguments{
}
\description{
wiki_graph
}
\references{
\url{https://en.wikipedia.org/wiki/Graph}
}
\keyword{datasets}
