% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wrap_cvodes}
\alias{wrap_cvodes}
\title{Returned is the Jacobian matrix as an \code{arma::mat} from the
    Armadillo package.}
\usage{
wrap_cvodes(times, states_, parameters_, initSens_, events_, settings, model_,
  jacobian_, sens_)
}
\value{
Matrix with nrow = (no. timepoints) and
    ncol = (no. states + no. observed + [(no. states)x(no. parameters)]).
    [(no. states)x(no. parameters)] is only returned if sensitivity equations
    are calculated.

    \describe{
    \item{First column}{
    Integration time points as given in \option{times}.}

    \item{Column 2 to no. of states + 1}{
    The state for the respective time point.}

    \item{no. of states + 1 to number of states + 1 + n. of observations}{
    Observation for the respective time point.}
    }
}
\description{
The list of arguments is the same as for \option{model_}.
}
\details{
This function sets up the cvodes integrator and loop over the
    vector \option{times} of requested time points. On success, the states
    of the system are returend for these time points.

    Write something about these observations, once you got a hold of them.
}
\author{
Alejandro Morales, \email{morales.s.alejandro@gmail.com}
}
