% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cOde.R
\name{writeOdeSystemSundials}
\alias{writeOdeSystemSundials}
\title{Implement the ODE system for sundials::cvodes.}
\usage{
writeOdeSystemSundials(f, variables, parameters, jac, not.zero.jac, filename)
}
\arguments{
\item{f}{Named character vector containing the right-hand sides of the ODE.
You may use the key word.}

\item{variables}{Variables appearing on the ODE, \code{names(f)}.}

\item{parameters}{Parameters appearing in the ODE.}

\item{filename}{Output file for c++ source code.}
}
\description{
Implement the ODE system for sundials::cvodes.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

