% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cOde.R
\name{cvodesSyntax}
\alias{cvodesSyntax}
\title{Sundials::cvodes: Replace powers and symbols to get correct C++ syntax for
cvodes.}
\usage{
cvodesSyntax(f, variables, parameters)
}
\arguments{
\item{f}{Named character vector containing the right-hand sides of the ODE.
You may use the key word.}

\item{variables}{Variables appearing on the ODE, \code{names(f)}.}

\item{parameters}{Parameters appearing in the ODE.}
}
\description{
Sundials::cvodes: Replace powers and symbols to get correct C++ syntax for
cvodes.
}
\author{
Wolfgang Mader, \email{Wolfgang.Mader@fdm.uni-freiburg.de}
}

