% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ensemble.R
\name{format_ensemble}
\alias{format_ensemble}
\title{Format ensemble}
\usage{
format_ensemble(ensemble, forecast_date, temporal_resolution = "wk")
}
\description{
Converts a quantile-only ensemble based on \code{\link[covidHubUtils:load_forecasts]{covidHubUtils::load_forecasts()}}
to standardised submission format
}
\details{
Steps:
Creates "target" variable
}
