% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.simulation.r
\name{module.simulation}
\alias{module.simulation}
\title{Module that performs the simulations by projections at the desired age starting from initial conditions.}
\usage{
module.simulation(stand = NULL)
}
\arguments{
\item{stand}{List with information originated from \code{\link{module.input}} containing:
SI, AGE0, HDOM0, BA0, N0, QD0, SDIR0, VOL_0B0, VOL_IB0, VOLm_0B0, VOLm_IB0,
AGEF, THINNING, AGET, BAR, t, d, method.}
}
\value{
A list containing a data frame (sim.stand) containing the columns below with simulations from
AGE0 to AGEF and a logical value indicating if thinning was requested.

\code{- AGE}     Stand age form simulations from AGE0 to AGEF (years).

\code{- HDOM}    Dominant Height (m).

\code{- SI}      Site Index (m).

\code{- BA}      Basal Area (m2/ha).

\code{- N}       Number of trees per hectare.

\code{- QD}      Mean quadratic diameter (cm).

\code{- SDIR}    Relative stand density index (\%).

\code{- VOL_OB}  Total stand-level volume outside bark (m3/ha).

\code{- VOL_IB}  Total stand-level volume inside bark (m3/ha).

\code{- VOLm_OB} Merchantable stand-level volume outside bark (m3/ha).

\code{- VOLm_IB} Merchantable stand-level volume inside bark (m3/ha).

\code{- CI}      Competition Index (required for THINNING=TRUE).

\code{- BAU}     Basal Area (m2/ha) of unthinned counterpart (required for THINNING
=TRUE).

\code{- NU}      Number of trees per hectare of unthinned counterpart (required for THINNING=TRUE).

}
\description{
\code{module.simulation} Performs model projections of a stand to a final age based on information provided
by the input module that contains:  SI, AGE0, HDOM0, BA0, N0, QD0, SDIR0, VOL_0B0, VOL_IB0,
VOLm_0B0, VOLm_IB0, AGEF, THINNING, AGET, BAR, t, d and method.
Provides with a data frame with all relevant information for simulations from AGE0 to AGEF and thinning (if requested).
}
\examples{
# Example 1 - Input from plot-level data (not thinning)
sim1 <- module.input(TYPE='PLOT', BA0=17.3, SI=30, AGE0=17, N0=1200, AGEF=28, THINNING=FALSE)
module.simulation(stand=sim1)

# Example 2 - Input from tree-level data
sim2<-module.input(TYPE='TREE', TREEDATA=treedata, AREA=500, AGE0=23, AGEF=32)
module.simulation(stand=sim2)

# Example 3 - Input from plot-level data (with thinning)
sim3 <- module.input(TYPE='PLOT', BA0=17.3, SI=30, AGE0=17, N0=1200, AGEF=38,
                     THINNING=TRUE, AGET=22, BAR=0.25)
sims <- module.simulation(stand=sim3)$sim.stand
sims
plot(sims$AGE,sims$BAU,type="l",xlim=c(10,50),ylim=c(10,50),col=2,
     xlab='Age (years)', ylab='Basal Area (m2/ha)')
par(new=TRUE)
plot(sims$AGE,sims$BA,type="l",xlim=c(10,50),ylim=c(10,50),col=1,
     xlab='Age (years)', ylab='Basal Area (m2/ha)')
par(new=FALSE)
}
\references{
Gonzalez-Benecke et al. (2012) - Modeling Survival, Yield, Volume Partitioning and Their Response to Thinning for Longleaf Pine Plantations.
Forests, 3(4), 1104-1132; doi:10.3390/f3041104
}
\seealso{
\code{\link{module.input}}
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
