% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ensemble.R
\name{run_ensemble}
\alias{run_ensemble}
\title{Run ensembling methods}
\usage{
run_ensemble(
  method = "mean",
  forecast_date,
  exclude_models = NULL,
  return_criteria = TRUE,
  verbose = FALSE,
  exclude_designated_other = TRUE,
  ...
)
}
\arguments{
\item{method}{name of ensembling method}

\item{forecast_date}{date or character}

\item{exclude_models}{character : names of team-models to exclude by forecast_date}

\item{return_criteria}{logical : whether to return a model/inclusion criteria grid
as well as the ensemble forecast (default \code{TRUE})}
}
\value{
\itemize{
\item if \code{return_criteria = TRUE}, a list with the following elements
\itemize{
\item "ensemble" : tibble : a single ensemble forecast
\item "criteria": tibble : all candidate models against criteria
for inclusion in ensemble (all locations and horizons)
\item "forecast_date" : date : latest date
}
\item if \code{return_criteria = FALSE}, a tibble of a single ensemble forecast
}
}
\description{
Run ensembling methods
}
\details{
Used to create a single ensemble forecast.
Takes a forecast date and loads all forecasts for the preceding week, then:
Filters models according to criteria
Ensembles forecasts according to given method
Formats ensemble forecast
Returns ensemble forecast and optionally the criteria for inclusion
}
