% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_models.R
\name{score_models}
\alias{score_models}
\title{Score models}
\usage{
score_models(forecasts, report_date, restrict_weeks)
}
\arguments{
\item{forecasts}{\code{data.frame} containing all the forecasts to be summarised
as an ensemble.}
}
\description{
Score models
}
