% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.STAND.r
\name{tree.STAND}
\alias{tree.STAND}
\title{Calculates the basal area, number of trees and mean quadratic diameter for a plot based on tree-level data.}
\usage{
tree.STAND(DBH, AREA)
}
\arguments{
\item{DBH}{Vector of diameter at breast height (DBH, cm). Must be complete}

\item{AREA}{Numeric value of the size of the plot (m2).}
}
\value{
A list containing the parameters:

\code{- BA} Basal Area (m2/ha).

\code{- N}  Number of trees per hectare.

\code{- QD} Mean Quadratic Diameter (cm).

}
\description{
\code{tree.STAND} Calculates the basal area, number of trees and mean quadratic diameter for a plot based on tree-level data.
}
\examples{
# Example - Stand information from inventory data
DBH <- treedata$DBH[!is.na(treedata$DBH)]
tree.STAND(DBH=DBH, AREA=301)
}
\author{
Priscila Someda-Dias, Salvador A. Gezan
}
