% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ensemble_criteria.R
\name{use_ensemble_criteria}
\alias{use_ensemble_criteria}
\title{Load models according to inclusion criteria before ensembling}
\usage{
use_ensemble_criteria(
  forecasts,
  exclude_models = NULL,
  exclude_designated_other = TRUE,
  return_criteria = TRUE
)
}
\arguments{
\item{forecasts}{\code{data.frame} containing all the forecasts to be summarised
as an ensemble.}

\item{exclude_models}{optional character vector to exclude over all dates,
or data.frame with cols model and forecast_date, to exclude for specific
dates}

\item{exclude_designated_other}{logical: whether to exclude models designated
as "other" in their metadata file (default \code{TRUE})}

\item{return_criteria}{logical : whether to return a model/inclusion criteria
grid as well as the ensemble forecast (default \code{TRUE})}
}
\value{
\itemize{
\item if \code{return_criteria = TRUE}, a list with the following elements
\itemize{
\item "ensemble" : tibble : a single ensemble forecast
\item "criteria": tibble : all candidate models against criteria
for inclusion in ensemble (all locations and horizons)
\item "forecast_date" : date : latest date
}
\item if \code{return_criteria = FALSE}, a tibble of a single ensemble forecast
}
}
\description{
Load models according to inclusion criteria before ensembling
}
\details{
Steps:
Currently, models included based on having:
\enumerate{
\item All quantiles
\item 4 horizons
\item Not manually specified for exclusion
\item Not the hub ensemble
}
}
